Flytter oppretting av medlemskapoppslag request objekt og respons til flex ut fra brukersporsmal endepunktet til egne klasser
